+++ test/smokes/misspell/ignore_array/badspell.rb
@@ -1,0 +1,4 @@
+def foo
+  infomation
+  recieve
+end