+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Misc:
 - Use Docker BuildKit [#1245](https://github.com/sider/runners/pull/1245)
 - Fix Kotlin file extensions [#1252](https://github.com/sider/runners/pull/1252)
 - **PHPMD** Support the comma-separated list options as array [#1253](https://github.com/sider/runners/pull/1253)
+- **Misspell** Support `ignore` option as array [#1255](https://github.com/sider/runners/pull/1255)
 
 ## 0.29.3